+++ JEBMath/JEBMath/Geometry/MatrixFunctions.hpp
@@ -1,11 +1,9 @@
 #ifndef JEBMATH_GEOMETRY_MATRIXFUCTIONS_HPP
 #define JEBMATH_GEOMETRY_MATRIXFUCTIONS_HPP
 
-#include "JEBMath/Graphics/Matrix.hpp"
-
 namespace JEBMath {
 
-template <typename T, size_t N>
+template <typename T, unsigned N>
 void transpose(Matrix<T, N>& m)
 {
     for (auto i = 0u; i < N; ++i)
@@ -15,7 +13,7 @@ void transpose(Matrix<T, N>& m)
     }
 }
 
-template <typename T, size_t N>
+template <typename T, unsigned N>
 Matrix<T, N> transposed(Matrix<T, N>& m)
 {
     Matrix<T, N> t;
@@ -27,7 +25,7 @@ Matrix<T, N> transposed(Matrix<T, N>& m)
     return t;
 }
 
-template <typename T, typename U, size_t N>
+template <typename T, typename U, unsigned N>
 auto multiplyTransposed(const Matrix<T, N>& a, const Matrix<U, N>& b)
     -> Matrix<decltype(T() * U()), N>
 {
