+++ JEBMath/JEBMath/Geometry/MatrixOperators.hpp
@@ -2,12 +2,11 @@
 #define JEBMATH_GEOMETRY_MATRIXOPERATORS_HPP
 
 #include <ostream>
-#include "JEBMath/Graphics/Matrix.hpp"
-#include "JEBMath/Graphics/Vector.hpp"
+#include "Vector.hpp"
 
 namespace JEBMath {
 
-template <typename T, typename U, size_t N>
+template <typename T, typename U, unsigned N>
 bool operator==(const Matrix<T, N>& a, const Matrix<U, N>& b)
 {
     auto aData = a.data();
@@ -20,7 +19,7 @@ bool operator==(const Matrix<T, N>& a, const Matrix<U, N>& b)
     return true;
 }
 
-template <typename T, typename U, size_t N>
+template <typename T, typename U, unsigned N>
 Matrix<T, N>& operator+=(Matrix<T, N>& a, const Matrix<U, N>& b)
 {
     auto dataA = a.data();
@@ -32,7 +31,7 @@ Matrix<T, N>& operator+=(Matrix<T, N>& a, const Matrix<U, N>& b)
     return a;
 }
 
-template <typename T, typename U, size_t N>
+template <typename T, typename U, unsigned N>
 auto operator+(const Matrix<T, N>& a, const Matrix<U, N>& b)
     -> Matrix<decltype(T() + U()), N>
 {
@@ -41,7 +40,7 @@ auto operator+(const Matrix<T, N>& a, const Matrix<U, N>& b)
     return c += b;
 }
 
-template <typename T, typename U, size_t N>
+template <typename T, typename U, unsigned N>
 Matrix<T, N>& operator-=(Matrix<T, N>& a, const Matrix<U, N>& b)
 {
     auto dataA = a.data();
@@ -53,7 +52,7 @@ Matrix<T, N>& operator-=(Matrix<T, N>& a, const Matrix<U, N>& b)
     return a;
 }
 
-template <typename T, typename U, size_t N>
+template <typename T, typename U, unsigned N>
 auto operator-(const Matrix<T, N>& a, const Matrix<U, N>& b)
     -> Matrix<decltype(T() + U()), N>
 {
@@ -62,7 +61,7 @@ auto operator-(const Matrix<T, N>& a, const Matrix<U, N>& b)
     return c -= b;
 }
 
-template <typename T, typename U, size_t N>
+template <typename T, typename U, unsigned N>
 auto operator*(const Matrix<T, N>& a, const Matrix<U, N>& b)
     -> Matrix<decltype(T() * U()), N>
 {
@@ -81,13 +80,13 @@ auto operator*(const Matrix<T, N>& a, const Matrix<U, N>& b)
     return c;
 }
 
-template <typename T, typename U, size_t N>
+template <typename T, typename U, unsigned N>
 Matrix<T, N>& operator*=(Matrix<T, N>& a, const Matrix<U, N>& b)
 {
     return a = a * b;
 }
 
-template <typename T, typename U, size_t N>
+template <typename T, typename U, unsigned N>
 auto operator*(const Matrix<T, N>& m, const Vector<U, N>& v)
     -> Vector<decltype(T() * U()), N>
 {
@@ -105,7 +104,7 @@ auto operator*(const Matrix<T, N>& m, const Vector<U, N>& v)
     return result;
 }
 
-template <typename T, typename U, size_t N>
+template <typename T, typename U, unsigned N>
 auto operator*(const Vector<T, N>& v, const Matrix<U, N>& m)
     -> Vector<decltype(T() * U()), N>
 {
@@ -123,7 +122,7 @@ auto operator*(const Vector<T, N>& v, const Matrix<U, N>& m)
     return result;
 }
 
-template <typename T, size_t N>
+template <typename T, unsigned N>
 std::ostream& operator<<(std::ostream& os, const Matrix<T, N>& m)
 {
     os << m[0][0];
